# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/merchants/{id}": {
      "get": {
        "operationId": "AdminController_merchantDetail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Full tenant detail: health, programme, staff, activity",
        "tags": [
          "Admin · Panel"
        ]
      }
    },
...
    "/v1/merchant/customers": {
      "get": {
        "operationId": "MerchantCustomersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Matches name, phone digits or customer code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/MembershipListItemDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List/search customers of this business (paginated)",
        "tags": [
          "Merchant · Customers"
        ]
      }
    },
    "/v1/merchant/customers/{membershipId}": {
      "get": {
        "operationId": "MerchantCustomersController_detail",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Customer detail (card state + campaign)",
        "tags": [
          "Merchant · Customers"
        ]
      },
      "patch": {
        "operationId": "MerchantCustomersController_updateCustomer",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update private notes and tags",
        "tags": [
          "Merchant · Customers"
        ]
      }
    },
    "/v1/merchant/customers/{membershipId}/stamps": {
      "get": {
        "operationId": "MerchantCustomersController_stampHistory",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/StampDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stamp history for a customer (paginated, newest first)",
        "tags": [
          "Merchant · Customers"
        ]
      },
      "post": {
        "operationId": "MerchantCustomersController_addStamp",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddStampResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add one stamp as the owner (solo merchants without staff)",
        "tags": [
          "Merchant · Customers"
        ]
      }
    },
    "/v1/merchant/customers/{membershipId}/adjust": {
      "post": {
        "operationId": "MerchantCustomersController_adjust",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustBalanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Correct a balance by ±N with a mandatory reason (recorded in the ledger)",
        "tags": [
          "Merchant · Customers"
        ]
      }
    },
    "/v1/merchant/customers/{membershipId}/block": {
      "post": {
        "operationId": "MerchantCustomersController_block",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlockCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipDetailDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Block or unblock a customer. Blocked members keep their history.",
        "tags": [
          "Merchant · Customers"
        ]
      }
    },
    "/v1/merchant/customers/{membershipId}/consents": {
      "get": {
        "operationId": "MerchantCustomersController_consents",
        "parameters": [
          {
            "name": "membershipId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Consent history — what they agreed to and when",
        "tags": [
          "Merchant · Customers"
        ]
      }
    },
    "/v1/merchant/redemptions": {
      "get": {
        "operationId": "MerchantRedemptionsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by voucher status",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "REDEEMED",
                "VOID"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Customer name, phone digits, customer code or voucher code",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/PaginatedDto"
                    },
                    {
                      "properties": {
                        "items": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/RedemptionRowDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List reward vouchers (pending first, newest first)",
        "tags": [
          "Merchant · Rewards"
        ]
      }
    },
    "/v1/merchant/redemptions/redeem": {
      "post": {
        "operationId": "MerchantRedemptionsController_redeem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedeemResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a voucher as handed over, as the owner",
        "tags": [
          "Merchant · Rewards"
        ]
      }
    },
...
    "/v1/staff/redemptions/redeem": {
      "post": {
        "operationId": "StaffConsoleController_redeem",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeemRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedeemResultDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark a reward voucher as handed over (by id or code)",
        "tags": [
          "Staff console"
        ]
      }
    },
...
  },
  "info": {
    "title": "Loyalty Platform API",
    "description": "Multi-tenant loyalty platform — Phase 1. Bearer JWT auth; see /docs for the interactive UI.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RedeemRequestDto": {
        "type": "object",
        "properties": {
          "redemptionId": {
            "type": "string",
            "description": "Redemption id from a card/search payload"
          },
          "code": {
            "type": "string",
            "example": "RX7K-9QZ2",
            "description": "Voucher code as shown on the customer card"
          }
        }
      },
...
      "PaginatedDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          }
        },
        "required": [
          "total",
          "page",
          "pageSize",
          "totalPages"
        ]
      },
...
      "MembershipListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Membership id"
          },
          "code": {
            "type": "string",
            "example": "7F3K9QZP"
          },
          "formattedCode": {
            "type": "string",
            "example": "7F3K-9QZP"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Private merchant notes"
          },
          "tags": {
            "example": [
              "regular"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "blockedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "blockedReason": {
            "type": "string",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSummaryDto"
          },
          "stampCount": {
            "type": "number",
            "example": 4
          },
          "stampsRequired": {
            "type": "number",
            "example": 10
          },
          "completedCount": {
            "type": "number",
            "example": 1
          },
          "totalStamps": {
            "type": "number",
            "example": 14
          },
          "lastStampAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pendingRewards": {
            "description": "Rewards waiting to be handed over",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RedemptionSummaryDto"
            }
          }
        },
        "required": [
          "id",
          "code",
          "formattedCode",
          "notes",
          "tags",
          "blockedAt",
          "blockedReason",
          "customer",
          "stampCount",
          "stampsRequired",
          "completedCount",
          "totalStamps",
          "lastStampAt",
          "joinedAt",
          "pendingRewards"
        ]
      },
...
      "MembershipDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Membership id"
          },
          "code": {
            "type": "string",
            "example": "7F3K9QZP"
          },
          "formattedCode": {
            "type": "string",
            "example": "7F3K-9QZP"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Private merchant notes"
          },
          "tags": {
            "example": [
              "regular"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "blockedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "blockedReason": {
            "type": "string",
            "nullable": true
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSummaryDto"
          },
          "stampCount": {
            "type": "number",
            "example": 4
          },
          "stampsRequired": {
            "type": "number",
            "example": 10
          },
          "completedCount": {
            "type": "number",
            "example": 1
          },
          "totalStamps": {
            "type": "number",
            "example": 14
          },
          "lastStampAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pendingRewards": {
            "description": "Rewards waiting to be handed over",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RedemptionSummaryDto"
            }
          },
          "campaign": {
            "$ref": "#/components/schemas/CardCampaignDto"
          }
        },
        "required": [
          "id",
          "code",
          "formattedCode",
          "notes",
          "tags",
          "blockedAt",
          "blockedReason",
          "customer",
          "stampCount",
          "stampsRequired",
          "completedCount",
          "totalStamps",
          "lastStampAt",
          "joinedAt",
          "pendingRewards",
          "campaign"
        ]
      },
      "StampDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "delta": {
            "type": "number",
            "example": 1,
            "description": "Signed movement: −1 rows are undo reversals or removals"
          },
          "reason": {
            "type": "string",
            "nullable": true,
            "description": "Present on adjustments and undos"
          },
          "completedCard": {
            "type": "boolean",
            "description": "True when this stamp completed a card and earned the reward"
          },
          "issuerType": {
            "type": "string",
            "enum": [
              "STAFF",
              "MERCHANT",
              "ADJUSTMENT"
            ]
          },
          "issuerName": {
            "type": "string",
            "example": "Ravi Kumar",
            "description": "Staff name, or \"Owner\" for merchant-issued stamps"
          }
        },
        "required": [
          "id",
          "createdAt",
          "delta",
          "reason",
          "completedCard",
          "issuerType",
          "issuerName"
        ]
      },
      "AddStampResultDto": {
        "type": "object",
        "properties": {
          "card": {
            "description": "Card state after the stamp",
            "allOf": [
              {
                "$ref": "#/components/schemas/MembershipListItemDto"
              }
            ]
          },
          "rewardEarned": {
            "type": "boolean",
            "description": "True when this stamp completed the card"
          },
          "reward": {
            "type": "string",
            "example": "1 free coffee of your choice"
          },
          "stamp": {
            "$ref": "#/components/schemas/StampDto"
          },
          "redemption": {
            "nullable": true,
            "description": "Voucher minted when this stamp completed the card",
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/RedemptionSummaryDto"
              }
            ]
          }
        },
        "required": [
          "card",
          "rewardEarned",
          "reward",
          "stamp",
          "redemption"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string",
            "description": "Private notes, visible only to the business"
          },
          "tags": {
            "example": [
              "regular",
              "birthday-march"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AdjustBalanceDto": {
        "type": "object",
        "properties": {
          "delta": {
            "type": "number",
            "example": -2,
            "description": "Whole number, positive or negative"
          },
          "reason": {
            "type": "string",
            "example": "Stamped twice by mistake at the counter"
          }
        },
        "required": [
          "delta",
          "reason"
        ]
      },
      "BlockCustomerDto": {
        "type": "object",
        "properties": {
          "blocked": {
            "type": "boolean",
            "description": "True to block, false to unblock"
          },
          "reason": {
            "type": "string",
            "example": "Repeated abuse of the referral bonus"
          }
        },
        "required": [
          "blocked"
        ]
      },
      "RedemptionRowDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Redemption (voucher) id"
          },
          "code": {
            "type": "string",
            "example": "RX7K9QZ2"
          },
          "formattedCode": {
            "type": "string",
            "example": "RX7K-9QZ2"
          },
          "rewardText": {
            "type": "string",
            "example": "1 free coffee of your choice",
            "description": "Snapshotted at earn time"
          },
          "earnedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "REDEEMED",
              "VOID"
            ]
          },
          "redeemedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "redeemedBy": {
            "type": "string",
            "nullable": true,
            "example": "Ravi Kumar",
            "description": "Who honoured it (\"Owner\" for merchant)"
          },
          "voidedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time",
            "description": "Set when the voucher was voided (e.g. its stamp was undone)"
          },
          "membershipId": {
            "type": "string",
            "description": "Membership id (links to the customer)"
          },
          "customer": {
            "$ref": "#/components/schemas/CustomerSummaryDto"
          },
          "customerCode": {
            "type": "string",
            "example": "7F3K-9QZP",
            "description": "Customer code"
          }
        },
        "required": [
          "id",
          "code",
          "formattedCode",
          "rewardText",
          "earnedAt",
          "status",
          "redeemedAt",
          "redeemedBy",
          "voidedAt",
          "membershipId",
          "customer",
          "customerCode"
        ]
      },
...
    }
  }
}